Split GroovySourceIncrementalCompilationIntegrationTest
into different classes. For this I did split all
the extra methods from GroovySourceIncrementalCompilationIntegrationTest
and moved them to GroovyIncrementalCompilationIntegrationTest.
Moreover, I also did split AbstractSourceIncrementalCompilationIntegrationTest
into AbstractClassChangeIncrementalCompilationIntegrationTest
causing the split of the sub-classes as well.
